refactor(cloudflare): remove BaseYjsRoom, inline into WorkspaceRoom + DocumentRoom
No inheritance — each DO is fully self-contained. Simplifies the migration
to a single v1 with both new_sqlite_classes since nothing is deployed yet.
